Feature-task4: *Added abstract class Model
